moved Player issueOrder, toAttack, and toDefend implementation to HumanPlayerStrategy. Added utility class for custom sort. moved EPlayerTurnPhase to CommonTypes. Player now calls the issueOrder, toAttack, and toDefend methods of PlayerStrategy.
