+++ COMP 345 Project.vcxproj
@@ -23,6 +23,7 @@
     <ClCompile Include="Source\Cards\Cards.cpp" />
     <ClCompile Include="Source\CommandProcessorDriver.cpp" />
     <ClCompile Include="Source\CommandProcessor\CommandProcessor.cpp" />
+    <ClCompile Include="Source\Common\Utilities.cpp" />
     <ClCompile Include="Source\Engine\GameEngine.cpp" />
     <ClCompile Include="Source\GameLog\GameLogDriver.cpp" />
     <ClCompile Include="Source\GameLog\LoggingObserver.cpp" />
@@ -41,6 +42,7 @@
     <ClInclude Include="Source\CommandProcessor\CommandProcessor.h" />
     <ClInclude Include="Source\Common\CommonTypes.h" />
     <ClInclude Include="Source\Common\localization.h" />
+    <ClInclude Include="Source\Common\Utilities.h" />
     <ClInclude Include="Source\Engine\GameEngine.h" />
     <ClInclude Include="Source\GameLog\LoggingObserver.h" />
     <ClInclude Include="Source\Map\map.h" />
