+++ Source/Common/Utilities.h
@@ -1,0 +1,8 @@
+#pragma once
+
+class Territory;
+
+class Utilities {
+public:
+	static bool SortByArmySize(const Territory* lhs, const Territory* rhs);
+};
