+++ Source/Player/Player.h
@@ -21,24 +21,23 @@ class Command;
 class CommandProcessor;
 
 enum class EOrderType;
+enum class EPlayerTurnPhase;
 
 /*
 	Summary: The Player class represents the player in a game of warzone.
 */
 class Player {
 public:
-	enum class EPlayerTurnPhase {
-		DeployingArmies,
-		AdvancingArmies,
-		PlayingCards,
-		EndOfTurn
-	};
-
 	/* ctors */
 	Player();
 	Player(std::string inPlayerName);
+	Player(std::string inPlayerName, PlayerStrategies* inPS);
+	Player(PlayerStrategies* inPS);
 	Player(const Player& inPlayer);
 	~Player();
+
+	friend class PlayerStrategies;
+
 	/* op overrides */
 	bool operator==(const Player& inRHS) const;
 	bool operator!=(const Player& inRHS) const;
@@ -56,9 +55,9 @@ public:
 
 	const std::size_t getPlayerID() const;
 	// returns list of owned territories adjacent to enemy territories in order of owned army strength
-	std::vector<Territory*> toDefend();
+	void toDefend();
 	// returns list of adjacent enemy territories in order of army strength
-	std::vector<Territory*> toAttack();
+	void toAttack();
 
 	std::vector<Territory*>& getTerritoriesToDefend();
 	const std::vector<Territory*>& getTerritoriesToDefend() const;
@@ -68,16 +67,12 @@ public:
 	void AddTerritoryToAttack(Territory * inTerritoryToAttack);
 	void AddTerritoryToDefend(Territory* inTerritoryToDefend);
 	void SetTerritoriesToAttack(std::vector<Territory*> inTerritories);
-	void SetTerritoriesToDefend(std::vector<Territory*> inTerritories);	
+	void SetTerritoriesToDefend(std::vector<Territory*> inTerritories);
 
-	// Unreal Engine style standard for long function declarations
-	void issueOrder
-	(
-		const EOrderType inOrderType,
-		const std::vector<Player*> inPlayers,
-		const std::vector<Territory*> inTerritories,
-		const std::vector<int> inNumArmies
-	);
+	// divides up the issueingOrders phase for the player
+	void setPlayerTurnPhase(EPlayerTurnPhase inPhase);
+	void setPlayerTurnPhase(int inPhase);
+	EPlayerTurnPhase getPlayerTurnPhase() const;
 
 	void issueOrder();
 
@@ -96,10 +91,6 @@ public:
 
 	void setCommandProcessor(CommandProcessor* inProcessor);
 
-protected:
-	Hand* _hand = nullptr;
-	OrdersList* _orders = nullptr;
-
 	// helper functions for issuing orders to the orders list
 	void IssueDeployOrder(Territory* inDst, uint32 inArmiesToDeploy);
 	void IssueAdvanceOrder(Territory* inSrc, Territory* inDst, uint32 inArmiesToAdvance);
@@ -108,29 +99,11 @@ protected:
 	void IssueAirliftOrder(Territory* inSrc, Territory* inDst, std::size_t inArmiesToAirlift);
 	void IssueNegotiateOrder(Player* inTarget);
 
+protected:
+	Hand* _hand = nullptr;
+	OrdersList* _orders = nullptr;
+
 private:
-	// Human specific Player functions, requiring human input
-	// For specifying parameters for Orders to be issued
-	void DeployArmies_Human(int& inAvailableReserves);
-	void AdvanceArmies_Human();
-	void PlayingCards_Human();
-	void PlayingBombCard_Human();
-	void PlayingBlockadeCard_Human();
-	void PlayingAirliftCard_Human();
-	void PlayingDiplomacyCard_Human();
-
-	// divides up the issueingOrders phase for the player
-	void setPlayerTurnPhase(EPlayerTurnPhase inPhase);
-	void setPlayerTurnPhase(int inPhase);
-	EPlayerTurnPhase getPlayerTurnPhase() const;
-
-	// Helper functions for displaying territories to the human player
-	void DisplayToriesToDefendAndAdjacents();
-	void DisplayPlayerToriesToDefendAndAttack();
-	void DisplayPlayerToriesToDefend();
-	void DisplayPlayerToriesToAttack();
-
-	std::string GetUserInput(Command *& userCommand);
 	void HandleSaveEffect(Command* inCommand, std::string inMsg);
 
 	// tracks subphase of issueingOrders phase
